发送邮件

+++ app/Http/Controllers/UsersController.php
@@ -5,33 +5,19 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Auth;
+use Mail;
 
 class UsersController extends Controller
 {
     public function __construct()
     {
         $this->middleware('auth', [
-            'except' => ['show', 'create', 'store','index']
+            'except' => ['show', 'create', 'store','index','confirmEmail']
         ]);
         $this->middleware('guest', [
             'only' => ['create']
         ]);
     }
-    public function index()
-    {
-        $users = User::paginate(10);
-        return view('users.index', compact('users'));
-    }
-    public function create()
-    {
-        return view('users.create');
-    }
-
-    public function show(User $user)
-    {
-        return view('users.show', compact('user'));
-    }
-
     public function store(Request $request)
     {
         $this->validate($request, [
@@ -46,9 +32,48 @@ class UsersController extends Controller
             'password' => bcrypt($request->password),
         ]);
 
+        $this->sendEmailConfirmationTo($user);
+        session()->flash('success', '验证邮件已发送到你的注册邮箱上，请注意查收。');
+        return redirect('/');
+    }
+    public function confirmEmail($token)
+    {
+        $user = User::where('activation_token', $token)->firstOrFail();
+
+        $user->activated = true;
+        $user->activation_token = null;
+        $user->save();
+
         Auth::login($user);
-        session()->flash('success', '欢迎，您将在这里开启一段新的旅程~');
+        session()->flash('success', '恭喜你，激活成功！');
         return redirect()->route('users.show', [$user]);
+    }
+    protected function sendEmailConfirmationTo($user)
+    {
+        $view = 'emails.confirm';
+        $data = compact('user');
+        $from = 'summer@example.com';
+        $name = 'Summer';
+        $to = $user->email;
+        $subject = "感谢注册 Weibo 应用！请确认你的邮箱。";
+
+        Mail::send($view, $data, function ($message) use ($from, $name, $to, $subject) {
+            $message->from($from, $name)->to($to)->subject($subject);
+        });
+    }
+    public function index()
+    {
+        $users = User::paginate(10);
+        return view('users.index', compact('users'));
+    }
+    public function create()
+    {
+        return view('users.create');
+    }
+
+    public function show(User $user)
+    {
+        return view('users.show', compact('user'));
     }
     public function edit(User $user){
         $this->authorize('update',$user);
@@ -74,7 +99,7 @@ class UsersController extends Controller
         return redirect()->route('users.show', $user);
     }
     public function destroy(User $user){
-        $this->middleware('destroy',$user);
+        $this->authorize('destroy',$user);
         $user->delete();
         session()->flash('success','删除成功');
         return back();
